Added Previews.kt for Previews Functions... Corrected dradle files and dependencies...
