+++ compose_app/build.gradle
@@ -48,8 +48,9 @@ android {
 }
 
 dependencies {
-    implementation project(':data')
     implementation project(':compose_ui')
+    implementation project(':data')
+    implementation project(':domain')
 
     // Compose
     implementation 'androidx.core:core-ktx:1.7.0'
