+++ compose_app/src/main/java/com/compose_app/MainActivity.kt
@@ -7,7 +7,7 @@ import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import coil.annotation.ExperimentalCoilApi
 import com.compose_ui.Home
-import com.compose_ui.theme.NewsItemsAppTheme
+import com.compose_app.theme.NewsItemsAppTheme
 import data.koinModule.dataModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.GlobalContext.startKoin
