+++ compose_app/src/main/java/com/compose_app/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.compose_ui.theme
+package com.compose_app.theme
 
 import androidx.compose.ui.graphics.Color
