+++ compose_ui/src/main/java/com/compose_ui/Home.kt
@@ -1,6 +1,5 @@
 package com.compose_ui
 
-import android.content.res.Configuration
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -11,17 +10,13 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateListOf
-import androidx.compose.ui.tooling.preview.Preview
 import coil.annotation.ExperimentalCoilApi
 import coil.compose.rememberImagePainter
-import com.compose_ui.theme.NewsItemsAppTheme
 import domain.NewsItem
 import domain.NewsRepository
 import kotlinx.coroutines.flow.Flow
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
-import java.time.LocalDate
-import java.time.format.DateTimeFormatter
 
 @ExperimentalCoilApi
 class Home : KoinComponent {
@@ -44,19 +39,6 @@ class Home : KoinComponent {
     }
 
     @Composable
-    @Preview(name = "Light NewsItemsListPreview")
-    @Preview(
-        uiMode = Configuration.UI_MODE_NIGHT_YES,
-        showBackground = true,
-        name = "Dark NewsItemsListPreview"
-    )
-    fun NewsItemsListPreview() {
-        NewsItemsAppTheme {
-            NewsItemsList()
-        }
-    }
-
-    @Composable
     fun NewsItem(item: NewsItem) {
         Row {
             Image(painter = rememberImagePainter(item.context),
@@ -65,24 +47,6 @@ class Home : KoinComponent {
                 Text(text = item.shortHeadline!!)
                 Text(text = item.altText!!)
             }
-        }
-    }
-
-    @Composable
-    @Preview(name = "Light NewsItemPreview")
-    @Preview(
-        uiMode = Configuration.UI_MODE_NIGHT_YES,
-        showBackground = true,
-        name = "Dark NewsItemPreview"
-    )
-    fun NewsItemPreview() {
-        val item = domain.NewsItem("/contents/1803102",
-            "UWCL",
-            LocalDate.parse("1639947600", DateTimeFormatter.ISO_OFFSET_DATE_TIME),
-            "https://images.beinsports.com/lbC99KH5kQB05f29Mn7IpDMKiVM=/800x450/4080544-2021-12-15T195957Z_1016770655_UP1EHCF1JJVC0_RTRMADP_3_SOCCER-CHAMPIONS-FCB-KOG-REPORT.JPG",
-        "Women's Champions League draw")
-        NewsItemsAppTheme {
-            NewsItem(item)
         }
     }
